+++ CHANGES.rst
@@ -6,6 +6,9 @@ For changes before version 3.0, see ``HISTORY.rst``.
 5.3 (unreleased)
 ----------------
 
+- Provide ``AccessControl.get_safe_globals`` to facilitate safe use.
+
+
 
 5.2 (2021-07-30)
 ----------------
@@ -16,8 +19,6 @@ For changes before version 3.0, see ``HISTORY.rst``.
 5.1 (2021-07-30)
 ----------------
 NOTE: This release has been yanked from PyPI due to wheel build issues.
-
-- Provide ``AccessControl.get_safe_globals`` to facilitate safe use.
 
 - Fix a remote code execution issue by preventing access to
   ``string.Formatter`` from restricted code.